Refactor v1 handlers
